functions/asset_forecast/requirements.txt
pandas
functions-framework
google-cloud-bigquery
google-cloud-logging
flask
db_dtypes

# TODO: Remove dev requirements
openpyxl
black
pylint